fix settings model
